pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kanban"
version = "0.0.0"
description = "Kanban CLI (minimal scaffold)"
readme = ""
requires-python = ">=3.11"
dependencies = [
    "typer>=0.7.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.18.0",
    "pydantic-settings>=2.0.0",
    "authlib>=1.0.0",
    "cryptography>=40.0.0",
    "anyio>=3.0.0",
]

[tool.setuptools.packages.find]
include = ["kanban*"]

[project.scripts]
kanban = "kanban.cli.main:main"

[tool.pytest.ini_options]
asyncio_mode = "auto"
